pyproject.toml
[tool.poetry]
name = "workbotv2"
version = "0.1.0"
description = ""
authors = ["beetlejaws <pavelsarych@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
environs = "^14.2.0"
asyncio = "^3.4.3"
aiogram = "^3.21.0"
alembic = "^1.16.4"
sqlalchemy = "^2.0.42"
asyncpg = "^0.30.0"
aiogoogle = "^5.15.0"
aiohttp = "^3.12.15"
google-auth = "^2.40.3"
requests = "^2.32.4"
aiogram-dialog = "^2.4.0"
pytz = "^2025.2"
nats-py = "^2.11.0"
apscheduler = "^3.11.0"
ormsgpack = "^1.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
